Extract clean pages protection logic to its own class
