+++ modules/core/src/main/resources/META-INF/classnames.properties
@@ -1106,7 +1106,7 @@ org.apache.ignite.internal.processors.cache.persistence.freelist.CorruptedFreeLi
 org.apache.ignite.internal.processors.cache.persistence.migration.UpgradePendingTreeToPerPartitionTask
 org.apache.ignite.internal.processors.cache.persistence.pagemem.PageMemoryImpl$Segment
 org.apache.ignite.internal.processors.cache.persistence.pagemem.PageMemoryImpl$ThrottlingPolicy
-org.apache.ignite.internal.processors.cache.persistence.pagemem.PagesWriteSpeedBasedThrottle$ThrottleMode
+org.apache.ignite.internal.processors.cache.persistence.pagemem.CleanPagesProtectionThrottle$ThrottleMode
 org.apache.ignite.internal.processors.cache.persistence.snapshot.SnapshotDiscoveryMessage
 org.apache.ignite.internal.processors.cache.persistence.snapshot.SnapshotOperation
 org.apache.ignite.internal.processors.cache.persistence.snapshot.TrackingPageIsCorruptedException
